Implement adapter and repository for fetching and storing Cities
Integration tests need to be run via 'mvn verify' since Flyway doesn't
seem to work with DuckDb when ran programmatically. I think the reason
is it doesn't care about connection limit, while DuckDB is very strict
about it.

* Moved to the newest Spring version and JDK 21.
* Added CityInformationClient together with @Scheduled refresh
(this won't work when running the application as the URL is a fake one).
* Added Wiremock and stubbed europeancitiesdictionary responses.
* Added few cities to the initial db script and added FK constraint.